pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "footprint"
version = "0.1.0"
description = "AI-powered travel journal pipeline"
requires-python = ">=3.11"
dependencies = [
    # Google
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.126.0",

    # Media / EXIF
    "Pillow>=10.3.0",
    "piexif>=1.1.3",
    "exifread>=3.0.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.29",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.3.2",
    "alembic>=1.13.1",

    # AI / Anthropic
    "anthropic>=0.26.0",

    # Config / env
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",

    # HTTP / async
    "httpx>=0.27.0",
    "aiofiles>=23.2.1",
    "tenacity>=8.3.0",

    # Geo
    "geopy>=2.4.1",

    # UI
    "streamlit>=1.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.4",
    "mypy>=1.10.0",
    "pip-tools>=7.4.1",
    "respx>=0.21.1",
]

[project.scripts]
footprint = "footprint.pipeline.runner:main"

[tool.hatch.build.targets.wheel]
packages = ["src/footprint"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
